# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/run:
    post:
      operationId: runBot
      tags: [bot]
      summary: Start bot run
      description: Start a new bot run with given config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotRunRequest"
      responses:
        "200":
          description: Bot run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotRunResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get bot run logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsResponse"
    delete:
      operationId: clearBotLogs
      tags: [bot]
      summary: Clear all bot logs
      responses:
        "200":
          description: Logs cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClearLogsResult"
  /bot/proxies:
    post:
      operationId: uploadProxies
      tags: [bot]
      summary: Upload proxy list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyUploadRequest"
      responses:
        "200":
          description: Proxies saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyUploadResponse"
...
components:
  schemas:
...
    BotRunRequest:
      type: object
      properties:
        referralCode:
          type: string
        accountCount:
          type: integer
          minimum: 1
        useProxy:
          type: boolean
      required:
        - referralCode
        - accountCount
        - useProxy
    BotRunResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        runId:
          type: string
      required:
        - success
        - message
...
    LogEntry:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
        level:
          type: string
          enum: [info, success, error, warning]
        message:
          type: string
        runId:
          type: string
      required:
        - id
        - timestamp
        - level
        - message
    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/LogEntry"
        total:
          type: integer
      required:
        - logs
        - total
    ClearLogsResult:
      type: object
      properties:
        success:
          type: boolean
        deleted:
          type: integer
        message:
          type: string
      required:
        - success
        - deleted
        - message
    ProxyUploadRequest:
      type: object
      properties:
        proxies:
          type: array
          items:
            type: string
      required:
        - proxies
    ProxyUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        message:
          type: string
      required:
        - success
        - count
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error